Handle invalid .clj files
In clojure 1.2, the reader only checks for invalid forms when the read forms are used,
so can throw outside of a read call. This commit forces any ns form to be read
completely befor ethe namespace name is exctracted.

See also http://dev.clojure.org/jira/browse/TNS-1 for discussion of the issue.