Added support for reader conditional files (cljc)
Clojure 1.7.0 will add support for reader conditionals which means that clojure tooling will soon need to be able to deal with namespaces defined in `cljc` files in addition to `clj` files. This patch adds support for resolving namespaces in `cljc` files while maintaining external API compatibility with the previous version.